+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 2.7.0
 -----
 
+ * added option "choice_translation_domain" to ChoiceType.
  * deprecated option "precision" in favor of "scale"
  * deprecated the overwriting of AbstractType::setDefaultOptions() in favor of overwriting AbstractType::configureOptions().
  * deprecated the overwriting of AbstractTypeExtension::setDefaultOptions() in favor of overwriting AbstractTypeExtension::configureOptions().
@@ -16,7 +17,7 @@ CHANGELOG
  * deprecated ChoiceToBooleanArrayTransformer and ChoicesToBooleanArrayTransformer
  * deprecated FixCheckboxInputListener and FixRadioInputListener
  * deprecated the "choice_list" option of ChoiceType
- * added new options to ChoiceType: 
+ * added new options to ChoiceType:
    * "choices_as_values"
    * "choice_loader"
    * "choice_label"